ENHANCEMENT: merge utils.go & main.go (#18)
